+++ components/base/base_test.go
@@ -13,7 +13,7 @@ import (
 	_ "go.viam.com/rdk/components/base/fake"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
-	robotimpl "go.viam.com/rdk/robot/impl"
+	robotimpltest "go.viam.com/rdk/robot/impltest"
 	"go.viam.com/rdk/testutils"
 	"go.viam.com/rdk/testutils/inject"
 )
@@ -101,7 +101,7 @@ func TestFromRobot(t *testing.T) {
 
 	conf := testutils.ConfigFromJSON(t, jsonData)
 	logger := logging.NewTestLogger(t)
-	r := robotimpl.SetupLocalRobot(t, context.Background(), conf, logger)
+	r := robotimpltest.SetupLocalRobot(t, context.Background(), conf, logger)
 
 	expected := []string{"base1", "base2"}
 	testutils.VerifySameElements(t, base.NamesFromRobot(r), expected)
