+++ components/board/board_test.go
@@ -10,7 +10,7 @@ import (
 	_ "go.viam.com/rdk/components/board/fake"
 	_ "go.viam.com/rdk/components/motor/fake"
 	"go.viam.com/rdk/logging"
-	robotimpl "go.viam.com/rdk/robot/impl"
+	robotimpltest "go.viam.com/rdk/robot/impltest"
 	"go.viam.com/rdk/testutils"
 )
 
@@ -47,7 +47,7 @@ func TestFromRobot(t *testing.T) {
 
 	conf := testutils.ConfigFromJSON(t, jsonData)
 	logger := logging.NewTestLogger(t)
-	r := robotimpl.SetupLocalRobot(t, context.Background(), conf, logger)
+	r := robotimpltest.SetupLocalRobot(t, context.Background(), conf, logger)
 
 	expected := []string{"board1"}
 	testutils.VerifySameElements(t, board.NamesFromRobot(r), expected)
