+++ components/motor/motor_test.go
@@ -14,7 +14,7 @@ import (
 	_ "go.viam.com/rdk/components/motor/fake"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
-	robotimpl "go.viam.com/rdk/robot/impl"
+	robotimpltest "go.viam.com/rdk/robot/impltest"
 	"go.viam.com/rdk/testutils"
 	"go.viam.com/rdk/testutils/inject"
 )
@@ -165,7 +165,7 @@ func TestFromRobot(t *testing.T) {
 
 	conf := testutils.ConfigFromJSON(t, jsonData)
 	logger := logging.NewTestLogger(t)
-	r := robotimpl.SetupLocalRobot(t, context.Background(), conf, logger)
+	r := robotimpltest.SetupLocalRobot(t, context.Background(), conf, logger)
 
 	expected := []string{"m1", "m2"}
 	testutils.VerifySameElements(t, motor.NamesFromRobot(r), expected)
