+++ robot/impl/utils.go
@@ -28,15 +28,3 @@ func setupLocalRobot(
 	})
 	return r
 }
-
-// SetupLocalRobot create a local robot from a given config and ensure it is shutdown
-// after the test is complete. This includes waiting for any reconfiguration workers to
-// finish.
-func SetupLocalRobot(
-	t *testing.T,
-	ctx context.Context,
-	cfg *config.Config,
-	logger logging.Logger,
-) robot.LocalRobot {
-	return setupLocalRobot(t, ctx, cfg, logger)
-}
